+++ src/msx/core/msx_scc.h
@@ -10,7 +10,7 @@ struct MsxSccState {
     uint32_t sampleRate;
     uint32_t cpuClockHz;
     uint32_t stepScale;
-    uint64_t sampleAccumulator;
+    uint32_t sampleAccumulator;
     int32_t dcFilterX;
     int32_t dcFilterY;
     uint16_t ringReadIndex;
